dom/src/threads/test/test_newError.html

<html>
<head>
  <title>Test for DOM Worker Threads</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("newError_worker.js");

  worker.onmessage = function(event) {
    ok(false, "Shouldn't get a message!");
    SimpleTest.finish();
  }

  worker.onerror = function(event) {
    is(event.message, "foo!", "Got wrong error message!");
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
